Replace bower packages by meteor ones
Bower build process was sometimes unpredictable and had unspecified
dependencies (git) that caused bugs (eg #164, #172). Meteor packages
have better integration with the meteor build system and ensure
deterministic build, the downside being that we have to maintain
wrapper of JS package just for the purpose of publishing in
atmosphere. Others are also struggling with this, see for instance:

  https://github.com/MeteorCommunity/discussions/issues/14

Hopefully we'll have a better solution one day (official browserify
integration? Who knows.)

As a side effect the new release of perfect-scrollbar breaks the
style, this will be fixed in a later commit.

Fixes #172