Fix home header bar
The issue was introduced in 3b2eb0f but was only partially fixed (in
urgency) in 71b9a42. We sould have test to avoid these trivial
regressions! (I guess React will also yelp in this particular case by
removing the need to link the template and the "component", and thus
removing the possibility to break this link)

Fixes #434