Clean NPM scripts output
I've just discovered that the reason NPM outputs a giant "npm ERR!" message
asking to contact the author of the library and suggesting that it’s most likely
“not a problem with npm itself” when I run "npm run lint" and I forgot a
semicolon is because the return code of an NPM script has to be 0. So this
commit adds "exit 0;" at the end of the NPM commands.